Adding commandline options
Added a few commandline options . This references what appears to be an old nuget package that's not longer accessible from nuget.org. I might switch this out later on, but it something that was very easy to implement and used it before in other projects.

Added build script to help easily publish it.